refactored datastore memory impl
