+++ data-access/src/infrastructure-services-impl/datastore/memorydb/infrastructure/community/community.memory-repository.ts
@@ -1,9 +1,9 @@
-import { MemoryBaseAdapter } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-base-adapter";
-import { MemoryRepositoryBase } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-repository";
-import { DomainExecutionContext } from "../../../../app/domain/contexts/domain-execution-context";
-import { Community, CommunityProps } from "../../../../app/domain/contexts/community/community/community";
-import { CommunityRepository } from "../../../../app/domain/contexts/community/community/community.repository";
-import { UserEntityReference, UserProps } from "../../../../app/domain/contexts/user/user";
+import { MemoryBaseAdapter } from "../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-base-adapter";
+import { MemoryRepositoryBase } from "../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-repository";
+import { DomainExecutionContext } from "../../../../../app/domain/contexts/domain-execution-context";
+import { Community, CommunityProps } from "../../../../../app/domain/contexts/community/community/community";
+import { CommunityRepository } from "../../../../../app/domain/contexts/community/community/community.repository";
+import { UserEntityReference, UserProps } from "../../../../../app/domain/contexts/user/user";
 
 export class MemoryCommunity extends MemoryBaseAdapter implements CommunityProps {
   name: string;
