--- data-access/src/infrastructure-services-impl/datastore/memorydb/infrastructure/member.memory-uow.ts
@@ -1,11 +1,0 @@
-import { Member, MemberProps } from "../../../../app/domain/contexts/community/member/member";
-import { MemoryStore } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-store";
-import { MemoryUnitOfWork } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-unit-of-work";
-import { MemoryMemberRepository } from "./member.memory-repository";
-import { InProcEventBusInstance, NodeEventBusInstance } from "../../../../../seedwork/event-bus-seedwork-node";
-
-export const buildMemoryMemberUnitOfWork = (
-  memberMemoryStore: MemoryStore<MemberProps>,
-) => {
-  return new MemoryUnitOfWork(InProcEventBusInstance, NodeEventBusInstance, Member, memberMemoryStore, MemoryMemberRepository);
-};
