+++ data-access/src/infrastructure-services-impl/datastore/memorydb/infrastructure/service/service.memory-repository.ts
@@ -1,9 +1,9 @@
-import { CommunityEntityReference, CommunityProps } from "../../../../app/domain/contexts/community/community/community";
-import { DomainExecutionContext } from "../../../../app/domain/contexts/domain-execution-context";
-import { Service, ServiceProps } from "../../../../app/domain/contexts/community/service/service";
-import { ServiceRepository } from "../../../../app/domain/contexts/community/service/service.repository";
-import { MemoryBaseAdapter } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-base-adapter";
-import { MemoryRepositoryBase } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-repository";
+import { CommunityEntityReference, CommunityProps } from "../../../../../app/domain/contexts/community/community/community";
+import { DomainExecutionContext } from "../../../../../app/domain/contexts/domain-execution-context";
+import { Service, ServiceProps } from "../../../../../app/domain/contexts/community/service/service";
+import { ServiceRepository } from "../../../../../app/domain/contexts/community/service/service.repository";
+import { MemoryBaseAdapter } from "../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-base-adapter";
+import { MemoryRepositoryBase } from "../../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-repository";
 
 class MemoryService extends MemoryBaseAdapter implements ServiceProps  {
