--- data-access/src/infrastructure-services-impl/datastore/memorydb/infrastructure/violation-ticket.memory-uow.ts
@@ -1,11 +1,0 @@
-import { ViolationTicket, ViolationTicketProps } from "../../../../app/domain/contexts/cases/violation-ticket/v1/violation-ticket";
-import { MemoryStore } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-store";
-import { MemoryUnitOfWork } from "../../../../../seedwork/services-seedwork-datastore-memorydb/infrastructure/memory-unit-of-work";
-import { InProcEventBusInstance, NodeEventBusInstance } from "../../../../../seedwork/event-bus-seedwork-node";
-import { MemoryViolationTicketRepository } from "./violation-ticket.memory-repository";
-
-export const buildMemoryViolationTicketUnitOfWork = (
-  violationTicketMemoryStore: MemoryStore<ViolationTicketProps>,
-) => {
-  return new MemoryUnitOfWork(InProcEventBusInstance, NodeEventBusInstance, ViolationTicket, violationTicketMemoryStore, MemoryViolationTicketRepository);
-};
